# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Knowledge Base API — used by the web app and AI agents
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface ChangePasswordInput {
  currentPassword: string;
  /** @minLength 8 */
  newPassword: string;
}

export type AuthUserRole = typeof AuthUserRole[keyof typeof AuthUserRole];


export const AuthUserRole = {
  admin: 'admin',
  editor: 'editor',
  user: 'user',
} as const;

export interface AuthUser {
  id: number;
  email: string;
  name: string;
  role: AuthUserRole;
}

export interface GroupSummary {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
}

export interface ArticleSummary {
  id: number;
  slug: string;
  title: string;
  updatedAt: string;
  createdAt: string;
  /** @nullable */
  updatedByName?: string | null;
  /** True when the article has group restrictions */
  isRestricted?: boolean;
  /** True when the current user can view the content */
  canAccess?: boolean;
  groups?: GroupSummary[];
}

export interface Article {
  id: number;
  slug: string;
  title: string;
  /** HTML content of the article */
  content: string;
  updatedAt: string;
  createdAt: string;
  /** @nullable */
  updatedByName?: string | null;
  isRestricted: boolean;
  canAccess: boolean;
  groups: GroupSummary[];
  backlinks?: ArticleSummary[];
}

export interface ArticleListResponse {
  articles: ArticleSummary[];
  total: number;
}

export interface ArticleInput {
  /** @minLength 1 */
  title: string;
  /** HTML content */
  content: string;
  groupIds?: number[];
}

export interface ArticleUpdate {
  /** @minLength 1 */
  title?: string;
  content?: string;
  groupIds?: number[];
}

export interface ArticleGroupsInput {
  groupIds: number[];
}

export interface ArticleStats {
  totalArticles: number;
  restrictedArticles: number;
  totalGroups: number;
  totalUsers: number;
  recentlyUpdated: ArticleSummary[];
  oldestUpdated?: ArticleSummary[];
}

export interface ImageUploadInput {
  /** Uploaded file name (multipart/form-data handled by server middleware) */
  filename: string;
}

export interface ImageUploadResponse {
  url: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  editor: 'editor',
  user: 'user',
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  createdAt: string;
  groups?: GroupSummary[];
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  admin: 'admin',
  editor: 'editor',
  user: 'user',
} as const;

export interface UserInput {
  email: string;
  /** @minLength 1 */
  name: string;
  /** @minLength 8 */
  password: string;
  role: UserInputRole;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  admin: 'admin',
  editor: 'editor',
  user: 'user',
} as const;

export interface UserUpdate {
  email?: string;
  /** @minLength 1 */
  name?: string;
  role?: UserUpdateRole;
  /** @minLength 8 */
  password?: string;
}

export interface Group {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
  memberCount?: number;
}

export interface GroupDetail {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
  members: User[];
}

export interface GroupInput {
  /** @minLength 1 */
  name: string;
  description?: string;
}

export interface GroupUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
}

export interface GroupMemberInput {
  userId: number;
}

export interface AdminImportInput {
  /** Uploaded zip file name (multipart/form-data handled by server middleware) */
  filename: string;
  overwrite?: boolean;
}

export interface AdminImportResponse {
  imported: number;
  skipped: number;
  errors: string[];
}

export type ListArticlesParams = {
/**
 * Filter articles by title/content substring
 */
search?: string;
sort?: ListArticlesSort;
order?: ListArticlesOrder;
limit?: number;
offset?: number;
};

export type ListArticlesSort = typeof ListArticlesSort[keyof typeof ListArticlesSort];


export const ListArticlesSort = {
  title: 'title',
  updated_at: 'updated_at',
  created_at: 'created_at',
} as const;

export type ListArticlesOrder = typeof ListArticlesOrder[keyof typeof ListArticlesOrder];


export const ListArticlesOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;

export type ListArticlesMaintenanceParams = {
limit?: number;
offset?: number;
};