# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Knowledge Base API — used by the web app and AI agents
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /articles:
    get:
      operationId: listArticles
      tags: [articles]
      summary: List all articles (titles only for access-restricted ones)
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Filter articles by title/content substring
        - name: sort
          in: query
          schema:
            type: string
            enum: [title, updated_at, created_at]
            default: title
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Article list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleListResponse"
    post:
      operationId: createArticle
      tags: [articles]
      summary: Create a new article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Slug already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /articles/maintenance:
    get:
      operationId: listArticlesMaintenance
      tags: [articles]
      summary: List articles sorted by oldest last-modified (maintenance view)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Articles sorted by oldest updated_at first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleListResponse"
...
  /articles/{slug}:
    get:
      operationId: getArticle
      tags: [articles]
      summary: Get a single article by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Article with content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "403":
          description: Access restricted — user is not a member of the required group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateArticle
      tags: [articles]
      summary: Update an article
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleUpdate"
      responses:
        "200":
          description: Updated article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteArticle
      tags: [articles]
      summary: Delete an article
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /articles/{slug}/backlinks:
    get:
      operationId: getArticleBacklinks
      tags: [articles]
      summary: Get all articles that link to this article
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of articles linking to this one
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArticleSummary"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /articles/{slug}/export/md:
    get:
      operationId: exportArticleMarkdown
      tags: [articles]
      summary: Export article as Markdown file
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Markdown file
          content:
            text/markdown:
              schema:
                type: string
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /articles/{slug}/export/pdf:
    get:
      operationId: exportArticlePdf
      tags: [articles]
      summary: Export article as PDF
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /articles/{slug}/groups:
    put:
      operationId: setArticleGroups
      tags: [articles]
      summary: Set the groups that can access this article
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleGroupsInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # ─── Auth ──────────────────────────────────────────────────────────────────
...
    ArticleSummary:
      type: object
      required: [id, slug, title, updatedAt, createdAt]
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedByName:
          type: ["string", "null"]
        isRestricted:
          type: boolean
          description: True when the article has group restrictions
        canAccess:
          type: boolean
          description: True when the current user can view the content
        groups:
          type: array
          items:
            $ref: "#/components/schemas/GroupSummary"
    Article:
      type: object
      required: [id, slug, title, content, updatedAt, createdAt, groups, isRestricted, canAccess]
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        content:
          type: string
          description: HTML content of the article
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedByName:
          type: ["string", "null"]
        isRestricted:
          type: boolean
        canAccess:
          type: boolean
        groups:
          type: array
          items:
            $ref: "#/components/schemas/GroupSummary"
        backlinks:
          type: array
          items:
            $ref: "#/components/schemas/ArticleSummary"
    ArticleListResponse:
      type: object
      required: [articles, total]
      properties:
        articles:
          type: array
          items:
            $ref: "#/components/schemas/ArticleSummary"
        total:
          type: integer
    ArticleInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
          description: HTML content
        groupIds:
          type: array
          items:
            type: integer
    ArticleUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
        groupIds:
          type: array
          items:
            type: integer
    ArticleGroupsInput:
      type: object
      required: [groupIds]
      properties:
        groupIds:
          type: array
          items:
            type: integer
...
    GroupSummary:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
...