# lib/api-zod/src/generated/types/groupSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Knowledge Base API — used by the web app and AI agents
 * OpenAPI spec version: 0.1.0
 */

export interface GroupSummary {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
}